BLAN-833 - renamed events/response_error file to errors - renamed events/response_success file to success
